feat: add tests for check_raw_df
